putting locals storage in a separate file
